[feat] #107 에러 코드 추가

+++ module-domain/domain-inventories/src/main/java/com/chaing/domain/inventories/service/InventoryService.java
@@ -350,6 +350,9 @@ public class InventoryService {
     }
 
     public void disposalInventory(DisposalRequest request) {
+        if (request.actorType() == null || request.actorType().isBlank()) {
+            throw new InventoriesException(InventoriesErrorCode.INVALID_LOCATION_TYPE);
+        }
         String actorType = request.actorType().toUpperCase();
         if (actorType.equals("HQ")) {
             hqInventoryRepository.deleteByInventoryIdIn(request.inventoryIds());
